requirements.txt
Django==3.2.12
django-environ==0.8.1
